[category-main] useMemo 모듈화, 로딩 추가
